Moved old optic types into classic, removed OpticCan object
