+++ src/main/scala/json/JsonPath.scala
@@ -4,6 +4,7 @@ import json.Json._
 import json.PathElement.{Field, Index}
 import monocly._
 import monocly.functions.{Index => FIndex}
+import monocly.classic.{Iso, Optional}
 
 import scala.language.dynamics
