+++ src/main/scala/monocly/OpticCan.scala
@@ -1,22 +1,21 @@
 package monocly
 
-object OpticCan:
 
-  //        GetMany
-  //          /\
-  //         /  \
-  // GetOption  GetOneOrMore
-  //         \  /
-  //          \/
-  //          Get 
-  trait GetMany
-  trait GetOption extends GetMany
-  trait GetOneOrMore extends GetMany
-  trait Get extends GetOption with GetOneOrMore
+//        GetMany
+//          /\
+//         /  \
+// GetOption  GetOneOrMore
+//         \  /
+//          \/
+//          Get 
+trait GetMany
+trait GetOption extends GetMany
+trait GetOneOrMore extends GetMany
+trait Get extends GetOption with GetOneOrMore
 
-  //    Modify
-  //      ^
-  //      |
-  //  ReverseGet
-  trait Modify
-  trait ReverseGet extends Modify
+//    Modify
+//      ^
+//      |
+//  ReverseGet
+trait Modify
+trait ReverseGet extends Modify
