+++ src/main/scala/monocly/functions/Each.scala
@@ -3,4 +3,4 @@ package monocly.functions
 import monocly._
 
 abstract class Each[S, A]:
-  def each: Optic[OpticCan.GetMany & OpticCan.Modify, S, A]
+  def each: Optic[GetMany & Modify, S, A]
